+++ src/components/Affiliation/index.js
@@ -1,0 +1,3 @@
+import component from './Affiliation'
+
+export default component