+++ src/routes/PrimaryLayout/components/Drawer/Drawer.test.js
@@ -20,22 +20,29 @@ const networks = [
   }
 ]
 
+const match = {
+  match: {
+    params: 'c'
+  }
+}
+
 describe('Drawer', () => {
   it('renders with a current community', () => {
     const wrapper = shallow(<Drawer
       community={communities[0]}
       communities={communities}
-      networks={[]} />)
+      networks={[]}
+      match={match} />)
     expect(wrapper).toMatchSnapshot()
   })
 
   it('renders without a current community', () => {
-    const wrapper = shallow(<Drawer communities={communities} networks={[]} />)
+    const wrapper = shallow(<Drawer communities={communities} networks={[]} match={match} />)
     expect(wrapper).toMatchSnapshot()
   })
 
   it('renders a community list if networks are present', () => {
-    const wrapper = shallow(<Drawer communities={communities} networks={networks} />)
+    const wrapper = shallow(<Drawer communities={communities} networks={networks} match={match} />)
     expect(wrapper).toMatchSnapshot()
   })
 })