+++ src/routes/UserSettings/CommunitySettingsTab/index.js
@@ -1,3 +1,3 @@
 import CommunitySettingsTab from './CommunitySettingsTab'
-
-export default CommunitySettingsTab
+import connector from './CommunitySettingsTab.connector'
+export default connector(CommunitySettingsTab)