+++ src/routes/UserSettings/UserSettings.connector.js
@@ -3,31 +3,20 @@ import { push } from 'connected-react-router'
 import getMe from 'store/selectors/getMe'
 import {
   fetchSavedSearches, deleteSearch, viewSavedSearch,
-  updateUserSettings, leaveCommunity, unlinkAccount,
+  updateUserSettings, unlinkAccount,
   updateMembershipSettings, updateAllMemberships, registerStripeAccount
 } from './UserSettings.store'
 import { generateViewParams } from 'util/savedSearch'
 import { setConfirmBeforeClose } from '../FullPageModal/FullPageModal.store'
 import { loginWithService } from 'routes/NonAuthLayout/Login/Login.store'
-import { createSelector as ormCreateSelector } from 'redux-orm'
 import { createSelector } from 'reselect'
-import orm from 'store/models'
 import fetchForCurrentUser from 'store/actions/fetchForCurrentUser'
 import unBlockUser from 'store/actions/unBlockUser'
 import getBlockedUsers from 'store/selectors/getBlockedUsers'
+import getCurrentUserMemberships from 'store/selectors/getCurrentUserMemberships'
 import { FETCH_FOR_CURRENT_USER, FETCH_SAVED_SEARCHES } from 'store/constants'
 import { get, every, includes } from 'lodash/fp'
 import getQuerystringParam from 'store/selectors/getQuerystringParam'
-
-// this selector assumes that all Memberships belong to the currentUser
-// using this instead of currentUser.memberships to avoid a memoization issue
-// see https://github.com/tommikaikkonen/redux-orm/issues/117
-export const getCurrentUserMemberships = ormCreateSelector(
-  orm,
-  state => state.orm,
-  (session) =>
-    session.Membership.all().toModelArray()
-)
 
 export const getAllCommunitiesSettings = createSelector(
   getCurrentUserMemberships,
@@ -47,9 +36,9 @@ export const getMessageSettings = createSelector(
 
 export function mapStateToProps (state, props) {
   const currentUser = getMe(state, props)
-  const memberships = getCurrentUserMemberships(state, props)
   const blockedUsers = getBlockedUsers(state, props)
   const allCommunitiesSettings = getAllCommunitiesSettings(state, props)
+  const memberships = getCurrentUserMemberships(state, props)
   const messageSettings = getMessageSettings(state, props)
   const searches = state.SavedSearches.searches
 
@@ -61,11 +50,11 @@ export function mapStateToProps (state, props) {
 
   return {
     currentUser,
-    memberships,
     blockedUsers,
     confirm,
     fetchPending,
     allCommunitiesSettings,
+    memberships,
     messageSettings,
     searches,
     queryParams
@@ -79,7 +68,6 @@ export function mapDispatchToProps (dispatch) {
     deleteSearch: (params) => dispatch(deleteSearch(params)),
     updateUserSettings: (params) => dispatch(updateUserSettings(params)),
     unBlockUser: (params) => dispatch(unBlockUser(params)),
-    leaveCommunity: (params) => dispatch(leaveCommunity(params)),
     loginWithService: (params) => dispatch(loginWithService(params)),
     unlinkAccount: (params) => dispatch(unlinkAccount(params)),
     setConfirmBeforeClose: (params) => dispatch(setConfirmBeforeClose(params)),